+++ library/src/main/java/com/egeniq/androidtvprogramguide/entity/ProgramGuideChannel.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Egeniq
+ * Copyright (c) 2020, Egeniq
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,11 @@ package com.egeniq.androidtvprogramguide.entity
 
 import android.text.Spanned
 
-
+/**
+ * A channel which may be associated with multiple programmes.
+ * Channels are displayed on the left side of the screen, and display the image you have defined in the URL,
+ * and the name to the right of the image. ID is only used for identification purposes, and should be unique.
+ */
 interface ProgramGuideChannel {
     val id: String
     val name: Spanned?